Imported into Netbeans
Decided to import it into netbeans and manually move files to the git
repo. Not able to run things out of the box at the moment.
Different ways of detecting when the user has clicked a mine include:
a) Click a location, find out position of button, check if mine, end
game, display other mines and how correct user was with their flags.
b) Create different types of buttons (one for a mine, one for a blank,
one for a number).

Leaning toward method (a) since it'd be easier to implement and more
difficult for a user to get information from the buttons.

+Added temporary pictures
+Added a Button class
+Added the visual board (minefield still needs to be added)
+Added right click functionality